Generate null distributions of motif counts for each motif type

Generate null distributions of motif counts for each motif type

generate_nulls(
  edgelist = NULL,
  paralogs = NULL,
  edgelist_ppi = NULL,
  n = 1000,
  bp_ppiv = BiocParallel::SerialParam(),
  bp_v = BiocParallel::SerialParam(),
  bp_lambda = BiocParallel::SerialParam(),
  bp_delta = BiocParallel::SerialParam(),
  bp_bifan = BiocParallel::SerialParam()
)

Arguments

edgelist

A 2-column data frame with regulators in column 1 and targets in column 2.

paralogs

A 2-column data frame with gene IDs for each paralog in the paralog pair.

edgelist_ppi

A 2-column data frame with IDs of genes that encode each protein in the interacting pair.

n

Number of degree-preserving simulated networks to generate. Default: 1000.

bp_ppiv

BiocParallel back-end to be used in find_ppi_v(). Default: BiocParallel::SerialParam().

bp_v

BiocParallel back-end to be used in find_v(). Default: BiocParallel::SerialParam().

bp_lambda

BiocParallel back-end to be used in find_lambda(). Default: BiocParallel::SerialParam().

bp_delta

BiocParallel back-end to be used in find_delta(). Default: BiocParallel::SerialParam().

bp_bifan

BiocParallel back-end to be used in find_bifan(). Default: BiocParallel::SerialParam().

Value

A list of numeric vectors named lambda, delta, V, and bifan, containing the null distribution of motif counts for each motif type.

Examples

set.seed(123)
data(gma_grn)
data(gma_paralogs)
data(gma_ppi)
edgelist <- gma_grn[500:1000, 1:2] # reducing for test purposes
paralogs <- gma_paralogs[gma_paralogs$type == "WGD", 1:2]
edgelist_ppi <- gma_ppi
n <- 2 # small n for demonstration purposes
generate_nulls(edgelist, paralogs, edgelist_ppi, n)
#> $lambda
#> lambda lambda 
#>      0      0 
#> 
#> $delta
#> delta delta 
#>     0     0 
#> 
#> $V
#> V V 
#> 0 0 
#> 
#> $PPI_V
#> PPI_V PPI_V 
#>   190   186 
#> 
#> $bifan
#> bifan bifan 
#>     0     0 
#>